refactor: prettier-config 설정 적용 및 경로 수정
