+++ .eslintrc.js
@@ -1,0 +1,1 @@
+// module.exports = { extends: './eslint-config/common-ts' };
