+++ prettier-config/README.md
